Add option to merge tables rather than replacing
Add -add_tables option to buildifier, allowing users to merge their own
tables with the built-in tables rather than replacing the built-in
tables entirely. For users wanting to use this to format their own added
rules, this is much more useful than needing to duplicate the (mutable)
contents of the built-in tables.