Use sort.SliceStable instead of sort.Slice (#294)
In some cases the lexer adds comments with an empty string as a token to the tree to separate comment blocks inside a statement (e.g. if a list contains line comments inside). Such fake comments don't have real `Start.Byte` values, instead the value is the same as for the next real comment line.

Later, when the comments are sorted to restore the possibly broken order, such comments order can be broken by unstable sorting algorithms.