Expand the set of file names considered to be build files. (#400)
When using Bazel external workspaces, it's common to have files named like something.BUILD or BUILD.something, which become the BUILD file for the @something repository. This CL modifies build.ParseDefault so that it considers any filename with an extension or base called "BUILD" to be a BUILD file. While workspace file naming in the wild is less creative, I didn't see a reason not to extend the same rule to workspace files.