Change generate_tables from genrule to .bzl rule.
genrules do not work on Windows without a msys dependency.

The tool had to be changed as well, since ctx.actions.run does not
support piping stdin and stdout to files. I couldn't find any
other dependencies on the tool in this repository, but the visibility
is public so if any external dependencies have decided to use
this they will break.