fix(buildifier): default buildifier_test to "diff" mode
This commit changes the default value passed to the buildifier tool via
the `-mode` argument from `fix` to `diff` when defining a
`buildifier_test` target. This is necessary, as `fix` attempts to format
the files, which fails due to attempting to open and operator on files
in a read-only sandbox.

Given that `fix` is not an available value for the `mode` attribute to
the `buildifier_test` rule when manually specified, it makes sense to
remove it.

closes #971