Handle case when duplicate keys for the table column to be sorted.
When buildifier table sort for a column that has duplicate values,
then buildifier sorts the entries preserving the order of the duplicate keys
from the original table. Also added unitetests.

Unit testcase 81 - duplicate key for the col to be sorted for table sort tag, use case when comment appends multiple times in output
Unit testcase 82 - duplicate key for the col to be sorted for table sort tag - throws error