Add mailmap to correct author names and email addresses
This commit adds a .mailmap at the root of the repository in order to
map author names and email addresses to a canonical format (e.g.
combining duplicates, removing non-name strings). When different
email identities existed for the same author, the information from the
most recent commit made by the author was selected.

This plays nicely with `git-shortlog`, and `git-log` with the
`--use-mailmap` option (alternatively, specify `log.mailmap = True` in
your git configuration) [0].

[0]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-logmailmap