Enable buildozer commands containing newlines in command files.
buildozer can accept an argument for the set command with appropriately
escaped newlines to force multiline values that would otherwise be
written out as a single line. When buildozer reads commands from a
commands file, it reads a single line and attempts to apply it as a
command, so newlines can't be used to force a multiline value.

To enable forcing multiline values when using a commands file, this
updates buildozer to continue reading lines when a line ends with an
escaped newline to match the behavior on the command line. This
additionally adds a test case for using whitespace on the command-line
so that there is consistent coverage for the two modes.