Support bzlmod repositories
unused_deps seems to be broken for repositories that have migrated over
to bzlmod.

Specifically, the reference to the aspect must be done via the canonical
name for the workspace now.
(This probably even needs to change once WORKSPACE is defunct in Bazel 9)

This will break users of unused_deps if they are not on bzlmod but given
the latest release of Bazel 8 this is the direction we are headed.
Existing WORKSPACE users can pin to a previous version.
(Alternatively we could maybe support a flag)

fixes #1315