Add change command to buildozer to change rule kind in place
I'm actively working on a migration where I need to change the kind of a
rule from one to another in place.  Prior to this step, used move,
rename, and set to make the attributes of the two rule look comparable,
but now I need to switch their behavior by changing to the new rule.
Combing with multiple other commands in a single invocation allows for
doing reasonably complex changes in place.

I wasn't seeing a command to perform this operation, though I could have
missed that.  "rename" is already taken for attributes.  "change" seemed
the clearest after that. Open to other suggestions.

I've provided some basic test cases, and am happy to add more, but it
didn't seem particulary worthwhile, given the simplicity of the command
implementation.