enable_workspace when bazel verison > 8
If you don't want workspace enabled and are on version 8, unused_deps
did not work due to the lack of workspace support.

You could either fix this by adding it to your project's bazelrc however
we conditionally enable it here depending on the version of bazel being
executed.

This does have the side-effect that it seems to cause a rebuild due to
the arguments of the CLI changing.